requirements.txt
python-decouple==3.8
django==4.2.1
dj-rest-auth==4.0.1
django-ckeditor==6.5.1
djangorestframework==3.14.0
django-filter==23.2
psycopg2-binary==2.9.6
jdatetime==4.1.1
requests==2.31.0
Pillow==9.5.0
drf-dynamic-fields==0.4.0